Remove unused stuff from VaadinUtils
